@@ -2,13 +2,14 @@
   "name": "asset-build",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "scripts": {
     "build": "webpack build --mode production",
     "build-watch": "webpack watch --mode production",
     "build-dev": "webpack build --mode development",
     "build-dev-watch": "webpack watch --mode development",
-    "lint": "eslint --ext .js Resources/Private/Assets/JavaScript/",
-    "lint-fix": "eslint --fix --ext .js Resources/Private/Assets/JavaScript/",
+    "lint": "eslint Resources/Private/Assets/JavaScript/",
+    "lint-fix": "eslint --fix Resources/Private/Assets/JavaScript/",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -23,8 +24,7 @@
     "@babel/preset-env": "^7.25.0",
     "babel-loader": "^9.1.3",
     "core-js": "^3.37.1",
-    "eslint": "^8.57.0",
-    "eslint-webpack-plugin": "^4.2.0",
+    "eslint": "^9.8.0",
     "terser-webpack-plugin": "^5.3.10",
     "webpack": "^5.93.0",
     "webpack-cli": "^5.1.4"
@@ -34,5 +34,11 @@
     "Firefox ESR",
     "not ie <= 11",
     "not dead"
-  ]
+  ],
+  "devDependencies": {
+    "@eslint/eslintrc": "^3.1.0",
+    "@eslint/js": "^9.8.0",
+    "eslint-formatter-unix": "^8.40.0",
+    "globals": "^15.9.0"
+  }
 }
